MAINT: shuffle round robin head order
