+++ tests/data/round-robin.yaml
@@ -1,5 +1,5 @@
 # filename: round-robin.yaml
-targets: [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
+targets: ['vonn', 'moltzan', 'miller', 'ligety', 'raich', 'shiffrin', 'kildow', 'schlopy']
 html:
   max_queries: 20
 samplers:
